chore: update unit tests and assertions to support feature flag expiry in CDKv2 (#13217)
In CDKv2, we will be expiring the feature flags - 
'@aws-cdk/aws-ecr-assets:dockerIgnoreSupport' and 
'@aws-cdk/aws-s3:grantWriteWithoutAcl'.

Tests across the CDK code base explicitly or implicitly are verifying
the behaviour of these modules that are affected by these feature flags. 

Update the unit tests appropriately so that they explicitly depend on
one of the two behaviours, or test both.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*