fix(lambda-nodejs): esbuild detection with Yarn 2 in PnP mode
Refactor how `esbuild` is detected and run: check for a local version
first and fallback to a global installation.

For local bundling, if a local version is detected, run `esbuild` using
the right package manager.

In Docker, always run the globally installed version.

Support for PNPM could now easily be added in another PR.

Closes #13179