fix(cloudfront): cannot add two EdgeFunctions with same aliases
Attempting to add two EdgeFunctions to the same stack, and then giving both the
same alias (e.g., 'prod', 'live') results in an ID conflict. This is because the
aliases are being added to the scope of the shared EdgeFunction stack, rather
than the scope of the Function itself.

fixes #13237