fix(cloudfront): cannot add two EdgeFunctions with same aliases (#13324)
Attempting to add two EdgeFunctions to the same stack, and then giving both the
same alias (e.g., 'prod', 'live') results in an ID conflict. This is because the
aliases are being added to the scope of the shared EdgeFunction stack, rather
than the scope of the Function itself.

Note - Also took the opportunity to simplify one of the integ tests to deploy to a
single distribution.

fixes #13237


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*