chore: build go bindings for monocdk
Add `go` configuration to the `monocdk` package.

To be able to generate go bindings a few tweaks to the current codebase were needed:

BREAKING CHANGE: The deprecated `Duration.toISOString()` method is no longer available. Please use `Duration.toIsoString()` instead. Unfortunteally this API needs to be broken to disambiguate the two methods in the Go generated code.