chore: fix auto-approve workflow
I believe the reason our auto-approve workflow does not work is because it was triggered when the pull request was _created_ and before the `pr/auto-approve` label was applied to it.

This adds `types: [ labeled ]` so the workflow is triggered when a PR is labeled. The condition remains the same.