chore: PR linter GH action should run on edits and label adjusts
Currently, PR linter only runs when the PR is opened. This naturally
means that when the title, description or labels (that it complains
about) is fixed, it does not re-run and confirm that it passes.

Adjust the action so it also runs when the PR labels, title or
descriptions are modified.