fix(cfn-include): allow dynamic mappings to be used in Fn::FindInMap (#13428)
The template parsing logic in cloudformation-include always searched for the Mapping
in the template based on the first argument passed to Fn::FindInMap.
However, that doesn't work if that first argument is a dynamic expression,
like `{ Ref: Param }`.

Check for that case explicitly, and don't search for the Mapping
if the first argument to Fn::FindInMap is a dynamic expression.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*