fix(autoscaling): AutoScaling on percentile metrics doesn't work (#13366)
AutoScaling on percentile metrics did not work because the
`MetricAggregationType` was trying to be derived from the metric, and it can
only be MIN, MAX or AVG.

Figure out what the metric aggregation type does, default it to
AVERAGE if no other suitable value can be determined, and also make
it and the evaluation periods configurable while we're at it.

Fixes #13144.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*