chore(signer): mark aws-signer module as stable
CloudFormation supports only one resource type for AWS Signer, namely
[Signing Profile].

This is a very simple resource type with effectively two properties,
both of which, have been modeled according to the latest expectations of
the CDK construct library.

There is no good reason why this module ought to stay as experimental.
Given the standard nature of this module, I do not expect much customer
feedback on usability that would require backwards incompatible changes.
Hence, marking this as stable.

Motivation
The `ISigner` interface is referenced by the `Function` construct in the
`aws-lambda` module. This is an instance of a stable module depending on
an experimental module.

In CDKv2, we plan to ship only stable modules as part of the
mono-release of the cdk (aka, `aws-cdk-lib`). A prerequisite to achieve
this is to avoid such dependencies (stable → experimental).

In this specific case, we are fixing this by promoting the `aws-signer`
module to stable.

[Signing Profile]: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-signer-signingprofile.html