chore(ubergen): copy `.npmignore` into monopackage (#13642)
`.npmignore` did not use to be copied into the directories of the monopackage,
leading to files that *ought* to have been excluded to not be excluded.

Remove the file from the list of files that doesn't get copied to fix this.

This fix is not along the lines of the method suggested in the linked issue,
but it does address the original problem.

Fixes #12292.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*