chore(lambda-nodejs): update default runtime (#13664)
If the installed node version is >= 14, use Node.js 14.x. Otherwise, use
Node.js 12.x.

This removes a default on the deprecated Node.js 10.x runtime.

BREAKING CHANGE: the default runtime of a `NodejsFunction` is now Node.js 14.x if the environment from which it is deployed uses Node.js >= 14 and Node.js 12.x otherwise.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*