chore(ubergen): combine all Rosetta fixtures (#13670)
We need this to have `jsii-rosetta` runs that try to translate READMEs
from submodules not crash (because it can't find the referenced fixture
files).

This is not a correct final state yet, but it makes the build not crash.

Hopefully allows https://github.com/aws/jsii/pull/2712 to be merged.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*