chore(lambda-nodejs): prepare code to reduce merge conflicts when deprecated APIs are stripped (#13738)
BUNDLING - The `BundlingDockerImage` class is deprecated,
replaced by `DockerImage`. Remove all uses of this class
across modules in a backwards compatible way.

ECR ASSET - Change the way symbols are imported from
'@aws-cdk/assets' so that they can be in-place replaced with
their replacements from '@aws-cdk/core' module.
The `IAsset` interface from the core module introduces the
property `assetHash` instead of `sourceHash`. This change
makes the code forwards compatible so as to be supportive
on v2.

Other small backwards compatible changes.

BREAKING CHANGE: The type of `image` property in the
`Bundling` class is changed from `BundlingDockerImage` to
`DockerImage`.
* **lambda-nodejs**: The type of `dockerImage` property in
`BundlingOptions` is changed from `BundlingDockerImage` to
`DockerImage`.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*