chore(aws-cdk-lib): strip out deprecated symbols from public API (#13815)
The jsii build the 'aws-cdk-lib' module will now run with the 'strip
deprecated' flag enabled. This ensures that the public API of this
module will contain no deprecated symbols.

This is enabled by a new option configured in the module's
`package.json` and recognized by `cdk-build`.

Previous commits - ca391b59 and a872e672 - have removed the majority of
deprecated symbols from public APIs. A few remain that are removed as
part of this change.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*