chore: upgrade to netmask@2
netmask@1 is affected by CVE-2021-28918
https://github.com/advisories/GHSA-pch5-whg9-qr2r

netmask is a depdendency to the CDK via

  aws-cdk → proxy-agent@4.0.1 → pac-proxy-agent@4.1.0 →
  pac-resolver@4.1.0 → netmask@1.0.6

None of these dependencies have upgraded to netmask@2 as yet.

Use yarn's [selective dependency resolution], to explicitly pick
netmask@2. This upgrades yarn.lock and the CLI's npm-shrinkwrap.json.
With this fix, npm customers will no longer depend on netmask@2
transitively via the CDK.

For yarn customers, there is no clean resolution since yarn does not
respective the the 'resolutions' key in dependencies' package.json and
does not respect the shrinkwrap.
The init templates now ship the 'resolutions' key so that new customers
using yarn will be unaffected.

A different solution has to be devised for existing customers on yarn.

[selective dependency resolution]: https://classic.yarnpkg.com/en/docs/selective-version-resolutions/