chore: force shrinkwrap to accept incompatible versions
If the **resolved** version of a package (the one we force using Yarn)
does not match the **required** version of a package (the one in
`package.json`), then NPM ignores the resolved version and installs
the package from the `requires` list anyway, *even* if it's processing
a shrinkwrap file.

Unfortunately, sometimes we have to force this situation, such as with
the `netmask` issue, where the required version is declared at `^1.0.6`
but we need to force it to be resolved to `2.0.1`.

Put simply, if resolved does not match required:

* NPM assumes *resolved* is wrong and will install *required*.
* We need it to just install *resolved* and not think too much.

We achieve this by overwriting what's in *required* in the shrinkwrap
file if we encounter this situation.