chore: fail early on impossible shrinkwrap (#13899)
If the **resolved** version of a package (the one we force using Yarn)
does not match the **required** version of a package (the one in
`package.json`), then NPM ignores the resolved version and installs
the package from the `requires` list anyway, *even* if it's processing
a shrinkwrap file.

Unfortunately, sometimes we have to force this situation, such as with
the `netmask` issue, where the required version is declared at `^1.0.6`
but we need to force it to be resolved to `2.0.1`.

Detect this situation early and bail out, so that we don't try to ship something
that doesn't have the version resolution we expected to ship.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*