revert: fix(aws-ecs): "broken splunk-logging `tag`-option in fargate platform version 1.4 (#13882)" (#13892)
This reverts commit e9d9299.

Unfortunately the above commit leads to broken deployments that are caused by behaviors upstream (i.e. in the ECS backend):

> Resource handler returned message: "Invalid request provided: Create TaskDefinition: Log driver splunk disallows options: splunk-tag (Service: AmazonECS; Status Code: 400; Error Code: ClientException; Request ID: d13abe56-36fa-4e81-b980-bf83789d4d0d; Proxy
: null)" (RequestToken: 546a74e2-a2eb-ef75-43e7-231fe8927096, HandlerErrorCode: InvalidRequest)

This means however, that the buggy behavior mentioned in #13881 persists.
I am in communication with AWS support on this issue and will keep you posted.

Sorry to have caused that mess. I should have marked the PR #13882 as draft before everything was clear.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*