fix(aws-stepfunctions): allow tokens in state machine names
Name validation for state machines was added in #13387. The validation did not take into consideration that the name may contain an unresolved token. This change skips name validation if there is an unresolved token as it does not make sense to attempt to validate something that is unknown. This change also allows a few more special characters as they are allowed through the AWS Console.

fixes #13946