chore: un-jsii and bundle `@aws-cdk/yaml-cfn` (#13933)
The `yaml-cfn` library is used as an implementation detail of
`cloudformation-include`, `aws-cdk`, `decdk`, and the monopackages.

Un-jsii it and bundle it into the framework packages that need it.

Does not need to be accessed over jsii and does not need to be exposed
in the monopackages.

----

The previous version of this PR (#13699) had to be reverted (https://github.com/aws/aws-cdk/commit/b1ffd335b6c41a26c1f88db2fc5a739c4c18c7fe), because of a bug in NPM
that causes `.npmignore` files in symlinked, bundled dependencies to be ignored.

This in turn caused undesired `.ts` files to be bundled, which caused `ts-node`
to exit with an error.

Introduce a new `cdk-postpack` tool which will strip files that NPM *should* have
ignored from the generated tarball, and add some verification to the `pack.sh` script
to make sure we don't accidentally bundle `.ts` files.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*