chore(cli): change Java init templates to be less repetitive
The Java init templates were recently fixed to use `new` instead
of the builder, because the builder would instantiate the wrong class
(`Stack` instead of `MyStack`) (#13988).

However, the structure was changed to have 3 different `new` calls
in 3 different comment blocks.

Change it back to the old structure: there is one instantiation, and
uncommenting just passes or doesn't pass the `env` property. Except
this time we use `new` instead of the builder.