chore: deprecate `@aws-cdk/yaml-cfn`
`aws-cdk-lib` needs to have no dependencies (except for `constructs`),
and we don't want to support this package as part of its public API.
Therefore, it needs to be bundled into all packages that consume it.

NPM has a bug that makes it impossible to `bundledDependencies` a
package that is linked from the monorepo.

We tried to work around that bug but it blocked our build because the
`.tgz` is sometimes not available during the `postpack` phase (#13933).

Simplest way around this now is to just copy/paste the implementation
of `yaml-cfn` 3 times. Should we ever need to bundle monorepo
dependencies again in the future, we can try to revive the `postpack`
solution of #13933.