fix(apigateway): cannot remove first api key from usage plan (#13817)
The UsagePlanKey resource connects an ApiKey with a UsagePlan. The API
Gateway service does not allow more than one UsagePlanKey for any given
UsagePlan and ApiKey combination. For this reason, CloudFormation cannot
replace this resource without either the UsagePlan or ApiKey changing.

A feature was added back in Nov 2019 - 142bd0e2 - that allows multiple
UsagePlanKey resources. The above limitation was recognized and logical
id of the existing UsagePlanKey was retained.

However, this unintentionally caused the logical id of the UsagePlanKey
to be sensitive to order. That is, when the 'first' UsagePlanKey
resource is removed, the logical id of the what was the 'second'
UsagePlanKey is changed to be the logical id of what was the 'first'.
This change to the logical id is, again, disallowed.

To get out of this mess, we do two things -

1. introduce a feature flag that changes the default behaviour for all
new CDK apps.

2. for customers with existing CDK apps who are would want to remove
UsagePlanKey resource, introduce a 'overrideLogicalId' option that they
can manually configure with the existing logical id.

fixes #11876

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*