feat(codepipeline): introduce the Action abstract class
Add an officially supported class to the API of the CodePipeline module
that contains some logic common to all implementations of the `IAction` interface,
like dealing with variable namespaces, or methods like `onStateChange()`.

Make the experimental class in codepipeline-actions with the same name extend this new class,
and make it stable.