chore(aws-cdk-lib): temporarily turn off stripping deprecated symbols
The 'strip deprecated' functionality has a bug when a deprecated
interface is inherited by a non-deprecated interface and they both live
in different files.

The problem arises due to the 'import' statement being stripped ahead of
the heritage clause.

Temporarily turn off stripping deprecated symbols to unblock CI, until a
fix is available.