fix(rds): deploy fails with "SubnetGroup not found" (#13986)
As it turns out, RDS stores the name of the SubnetGroup in lowercase on the server-side.
The problem is that we weren't changing the case in CloudFormation,
and thus expressions like `{ Ref: SubnetGroup }` were using the wrong casing,
resulting in the RDS service failing with a 404 error when providing a name for the SubnetGroup.

Fixes #13976

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*