chore(aws-cdk): do not use tmp directory for init tests
For some unknown, odd reason, `fs.stat()`, `fs.access()`, `fs.exists()`
etc. cannot consistently detect the existence of init template files
when created within the OS' temp directory.

This manifests only when executed in the jsii/superchain image but not
on OSX.

Switch the base directory for the tests to within the repository.