chore: jest config inheritance should use package resolution, not filesystem paths (#14105)
Packages should not rely on the file system locations of other
packages in the source repository, just on the contract that is
established by the package manager (i.e., NPM or Yarn will create
symlinks that allow you to `require()` declared dependencies).

Doing otherwise is the death of hermetic builds.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*