fix(cfn-include): allow deploy-time values in Parameter substitutions in Fn::Sub expressions (#14068)
When a Token that resolved to an object, like `{ Ref }`,
was passed as the value to be substituted for a CloudFormation Parameter,
and that Parameter was used in an Fn::Sub expression,
it caused incorrect CloudFormation to be generated
(an Fn::Join as an argument to Fn::Sub, which is not allowed).
Check for that case explicitly, and move that substitution into the Fn::Sub map instead.

Fixes #14047

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*