fix(core): `toJsonString()` does not deal correctly with list tokens
For reasons explained in the previous PR, list tokens need to be run
through a custom resource. This custom resource gets added during token
resolution, but because that is run multiple times the side effect
of adding the custom resource gets executed multiple times.

Add a cache to make sure the second execution returns the first answer
as the first one.

Not a fantastic solution, but given the constraints of Lazies, I'm not
sure we can do any better.

Fixes #14088.