chore(rds): remove the use of testFutureBehavior for RDS_LOWERCASE_DB_IDENTIFIER feature flag
This is a backport of the fix needed during this forward merge #14114.

The use testFutureBehavior and testLegacyBehavior APIs are reservd for
feature flags that will be 'expired' in CDKv2. expired means that the
feature flag is blocked and the behaviour defaulted to as if it was
enabled.

The RDS_LOWERCASE_DB_IDENTIFIER feature flag cannot be expired in CDKv2.
If it was, existing customers will have to destroy their RDS database
instance and create a new one, in order to migrate to CDKv2.

Instead, this feature flag will continue to exist in CDKv2 but it will
default to 'true'.