chore: vendor in a fixed submodule of lerna to fix script times
Lerna scripts have been running slower and slower, because it is
using an exponential algorithm on every invocation to check our
dependency graph for cycles.

Some builds on the v2 branch have started timing out because of this
behavior.

Fix the exponential behavior and vendor the fix in as a tarball.