chore: make `pkglint` self-contained after build
Another PR in the interest of making builds self-contained and
deterministic (and therefore: cacheable):

Today, on every invocation, `pkglint` will read a file from the root of
the repository.

This forces *every* packge in the repository to declare an explicit
dependency on `../../../release.json`, for the dependency graph to
be correct and complete (because think about it: if that file is read
during every package build, then its contents affect the result of that
package build, and so it must be declared!)

Instead, we copy `release.json` into `pkglint` at its build time. Right
now, `pkglint` *still* depends on `release.json`, but since
`release.json` is now part of `pkglint`s build artifacts (and implicitly
versioned as part of that), every downstream package can suffice by
simply declaring a dependency on `pkglint`--which they all already have.