feat(ecs): add support for elastic inference accelerators in ECS task defintions (#13950)
This PR would enable users to attach inference accelerators to their ECS tasks (currently not supported by Fargate). It adds an optional `inferenceAccelerators` property to the EC2 Task Definition. It also adds `resourceRequirement` property to the Container definition to enable containers to refer to the inference accelerators provided in the task definition.  

Closes #12460 

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*