chore: unexpire some feature flags
Some feature flags should not be expired in CDKv2. "expired" implies
that the flags will be enabled and users will be unable to disable it.

The following feature flags must not be unexpired, since it makes it
virtually impossible, in some cases, to deploy the CDK application
after upgrading to CDKv2. Deployments may cause service interruption or
may require stack re-creation.

- @aws-cdk/aws-apigateway:usagePlanKeyOrderInsensitiveId
- @aws-cdk/aws-rds:lowercaseDbIdentifier
- @aws-cdk/core:stackRelativeExports

These flags will still be enabled by default, but in cases where
deployment is difficult or impossible, users are now able to disable
these feature flags.