chore: fix runtime info to sanitize dotnet jsii runtime versions (#14172)
As part of the CDK version reporting metadata, we include the jsii runtime
environment as a pseudo-construct.

For Node, Java, and Python, this produces reasonable results and is properly
interpreted and stored service-side (e.g., node.js/v12.21.0, Python/3.8.8,
Java/11.0.2); however, for .NET it's more complicated:
DotNet/5.0.3/.NETCoreApp,Version=v3.1/1.0.0.0.

This causes issues with the parsing of the prefix-encoded constructs server-
side. To simplify, this change sanitizes the input to a set of well-known
characters.

fixes #13754


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*