feat(codepipeline): introduce the Action abstract class (#14009)
Add an officially supported class to the API of the CodePipeline module
that contains some logic common to all implementations of the `IAction` interface,
like dealing with variable namespaces, or methods like `onStateChange()`.

There was previously a class like that in the codepipeline-actions module,
but it was marked experimental, and had some strong opinions on the construction way of the subclasses
(required passing the `actionProperties` through a `super()` call in the constructor,
which is not very flexible).

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*