chore(aws-cdk-lib): extract experimental L2s from aws-cdk-lib  (#14135)
This PR adds logic to extract the L2 of experimental modules from `aws-cdk-lib`.  The reason for keeping the L1s of experimental modules is twofold: 

1. Removing the L1 means that modules with an experimental L2 are in a disadvantage in comparison to modules that does not have an L2. 
2. `CfnInclude` requires the L1.

The main logic was added in `ubergen`. If the monolithic package includes the key `ubergen.stripExperimental` in its `package.json` file, experimental modules source files will not be copied, instead we  execute `cfn2ts` to regenerate the L1.

#### Testing

1. Added a post-package script -`verify-stripped-exp.js`, to `aws-cdk-lib`. The script will install the tarball and verify that only L1 files are includes in experimental modules. 

2. verified that the generated tarball can be used with a consumer application, and that experimental modules are working properly.

adding `do-not-merege` as some stable modules still depends on experimental modules, the work to remove these dependencies is in progress.

#### Unrelated callout

We found a bug in a util function copied from `ubergen`, this PR adds a fix to both code path, but clearly this copy-pasta is not a great mechanism. This kind of copied util scripts is prevalent across our repo and we don't yet have a mechnisam to share sure common runtime functionally between modules. Not sure what is the right solution here but wanted to call it out, in case someone has a fix/methodology I have missed.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*