chore: backport test changes while adjusting feature flags in v2 (#14233)
This is a backport of 81148f02.

The defaults for some feature flags in CDKv2 are changing from 'false'
to 'true'.

Some test are sensitive to this change. This change modifies such tests
to explicitly set the value of the feature flag so that its behavior is
consistent across the two versions of the CDK.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*