Relicense init-templates to MIT-0
Previously the template setup.py would presume an Apache-2.0 license.
Rather than presume this for developers, we're relicensing the template
code to the MIT No Attribution license so that developers can choose
their own license and know that they have no obligations to Amazon for
the code that came from the template.