chore: keep core documentation in sync between @aws-cdk/core and aws-cdk-lib
Users of CDKv1 use the @aws-cdk/core documentation as the landing point for
documentation on core concepts like stacks, durations, references, etc. For
CDKv2 users, the same landing point is aws-cdk-lib, which currently has very
limited documentation.

This change introduces a pkglint rule to keep a subset of the two READMEs in
sync. This was chosen over syncing the entire README as the titles, heading,
stability markings, etc. needed to be different between the two packages.