chore: validate yarn.lock contains all necessary dependencies
Yarn v1 doesn't repect --frozen-lockfile entirely when using yarn workspaces.
The net effect is that if a dependency in one of our packages is not in the
yarn.lock file, yarn will happily install any/latest version of that package,
even when --frozen-lockfile is provided.

This came up recently when we noticed that the wrong version of constructs was
specified in yarn.lock, causing builds to ignore the lock file and pull in the
latest.