chore: validate yarn.lock contains all necessary dependencies (#14382)
Yarn v1 doesn't repect --frozen-lockfile entirely when using yarn workspaces.
The net effect is that if a dependency in one of our packages is not in the
yarn.lock file, yarn will happily install any/latest version of that package,
even when --frozen-lockfile is provided.

This came up recently when we noticed that the wrong version of constructs was
specified in yarn.lock, causing builds to ignore the lock file and pull in the
latest.

Note - This *will* cause the v2 branch to fail currently, so I'll be manually forward-merging
this change once approved to v2.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*

(cherry picked from commit f789757cc65ff2e1b65568365e77386051f0673b)

# Conflicts:
#	package.json