chore(yarn-cling): fix hoisting behavior (#14380)
Our original package hoisting behavior in `yarn-cling` was naive, and
might leave dependencies in places where they would actually lead to a
version conflict, or leave dependencies in places where they end up
being unused.

Nothing bad happened because of this: `npm ls` detects when the
installed dependencies don't match the declared dependencies, and fails
our "init template tests". This has in fact happened right now, where
our hoisting behavior leaves the new `punycode` dependency in a place
where it's unused, and now our tests are breaking because `npm ls` is
unhappy with the dependency tree.

Fortunately, I happened to have to deal with this already while working
on `nozem`, and so I had fixes for the broken hoisting behavior lying
around already.

This fixes the bugs in the hoisting, and thereby also the build.


----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*