feat(route53): add support for parentHostedZoneName for CrossAccountZoneDelegationRecord (#14097)
feat(route53): add support for parentHostedZoneName for CrossAccountZoneDelegationRecord

1. Zone name lookup: 

The current implementation of `CrossAccountZoneDelegationRecord` requires users to pass the `parentHostedZoneId` which is not so easy since we cannot pass references cross account.

This implementation lets them pass the zone name and queries the route53 API to get the zoneId.

2. Delegation role name addition

The ARN of the delegation role cannot be passed around to stacks since they will have different environments (this is cross account delegation and CDK does not allow passing references across environments).

This means that users need to import the delegation role in the stack containing the child zone. Since it needs to be imported, the users would need the role ARN. Now this ARN can either be hard-coded or constructed. Hard-coding the ARN is not entirely feasible since that would require deploying the parent stack first, then making code change and then deploying child stack in pipelines.

With role name, the ARN can be easily constructed (refer readme for importing role) and it provides a better user experience.

----

*By submitting this pull request, I confirm that my contribution is made under the terms of the Apache-2.0 license*