chore: fix NodeJS init template V2 unit tests for new-style synthesis
Our NodeJS init templates for 'app' and 'lib' come with a unit test that verifies the synthesized Stack is empty.
However, when we switched to new-style synthesis by default in V2 in #14182,
that assertion is no longer true -
new-style synthesis always adds a Parameter and a Rule to the template.

Change the assertion to verify that the `Resources` section is empty instead.